split index.js and index.test.js on functional files
